Reference

Data types

PikaParser.MatchResult — Type

A shortcut for possibly failed match result index (that points into ParserState field matches.

PikaParser.Clause — Type

abstract type Clause{G}

Abstract type for all clauses that match a grammar with rule labels of type G.

Currently implemented clauses:

Satisfy

TakeN

Token

Tokens

Epsilon

Fail

Seq

First

NotFollowedBy

FollowedBy

OneOrMore

ZeroOrMore

Often it is better to use convenience functions for rule construction, such as seq or token; see flatten for details.

PikaParser.Epsilon — Type

struct Epsilon{G} <: PikaParser.Clause{G}

An always-succeeding epsilon match.

PikaParser.Fail — Type

struct Fail{G} <: PikaParser.Clause{G}

An always-failing match.

PikaParser.First — Type

struct First{G} <: PikaParser.Clause{G}

Match the first possibility of several matches. Empty First is equivalent to unconditional failure.

Fields

children::Vector

PikaParser.FollowedBy — Type

struct FollowedBy{G} <: PikaParser.Clause{G}

Zero-length match that succeeds if follow does match at the same position.

Fields

follow::Any

PikaParser.Grammar — Type

struct Grammar{G}

A representation of the grammar prepared for parsing.

Fields

names::Vector

Topologically sorted list of rule labels (non-terminals)

idx::Dict{G, Int64} where G

Mapping of rule labels to their indexes in names

clauses::Vector{PikaParser.Clause{Int64}}

Clauses of the grammar converted to integer labels (and again sorted topologically)

can_match_epsilon::Vector{Bool}

Flags for the rules being able to match on empty string unconditionally

seed_clauses::Vector{Vector{Int64}}

Which clauses get seeded upon matching of a clause

terminals::Vector{Int64}

A summarized list of grammar terminals that are checked against each input letter

PikaParser.Match — Type

struct Match

Internal match representation.

Fields

pos::Int64

Where the match started?

len::Int64

How long is the match?

option_idx::Int64

Which possibility did we match?

submatches::Vector{Int64}

Indexes to the vector of matches. This forms the edges in the match tree.

PikaParser.MemoKey — Type

struct MemoKey

Index into the memoization table.

Fields

clause::Int64

start_pos::Int64

PikaParser.MemoTable — Type

mutable struct SortedDict{PikaParser.MemoKey, Int64, Ord<:Base.Order.Ordering} <: AbstractDict{PikaParser.MemoKey, Int64}

Pikaparser memoization table.

PikaParser.NotFollowedBy — Type

struct NotFollowedBy{G} <: PikaParser.Clause{G}

Zero-length match that succeeds if reserved does not match at the same position.

Fields

reserved::Any

PikaParser.OneOrMore — Type

struct OneOrMore{G} <: PikaParser.Clause{G}

Greedily matches a sequence of matches, with at least 1 match.

Fields

item::Any

PikaParser.ParseResult — Type

struct ParseResult

Representation of the parser output.

Fields

memo::DataStructures.SortedDict{PikaParser.MemoKey, Int64}

Best matches of grammar rules for each position of the input

matches::Vector{PikaParser.Match}

Match tree (folded into a vector)

PikaParser.ParserState — Type

mutable struct ParserState{G, I}

Intermediate parsing state. The match tree is built in a vector of matches that grows during the matching, all match indexes point into this vector.

Fields

grammar::PikaParser.Grammar

memo::DataStructures.SortedDict{PikaParser.MemoKey, Int64}

q::DataStructures.SortedSet{Int64}

matches::Vector{PikaParser.Match}

input::Any

PikaParser.Satisfy — Type

struct Satisfy{G} <: PikaParser.Clause{G}

A single terminal. Matches a token from the input stream where the match function returns true.

Fields

match::Function

PikaParser.Seq — Type

struct Seq{G} <: PikaParser.Clause{G}

Sequence of matches. Empty Seq is equivalent to an always-succeeding empty match, as in Epsilon.

Fields

children::Vector

PikaParser.TakeN — Type

struct TakeN{G} <: PikaParser.Clause{G}

A single terminal. Given the input stream and a position in it, the match function returns the length of the match, or nothing if there's no match.

Fields

match::Function

PikaParser.Token — Type

struct Token{G} <: PikaParser.Clause{G}

A single token equal to match.

Fields

token::Any

PikaParser.Tokens — Type

struct Tokens{G} <: PikaParser.Clause{G}

A series of tokens equal to match.

Fields

tokens::Vector

PikaParser.TraverseNode — Type

mutable struct TraverseNode{G}

Part of intermediate tree traversing state.

Fields

parent_idx::Int64

parent_sub_idx::Int64

rule::Any

match::PikaParser.UserMatch

open::Bool

subvals::Vector

PikaParser.UserMatch — Type

struct UserMatch{G}

User-facing representation of a Match.

Fields

pos::Int64

Where the match started?

len::Int64

How long is the match?

submatches::Array{Tuple{Int64, G}, 1} where G

Indexes and rule labels of the matched submatches. This forms the edges in the match tree.

PikaParser.ZeroOrMore — Type

struct ZeroOrMore{G} <: PikaParser.Clause{G}

Greedily matches a sequence of matches that can be empty.

Fields

item::Any

Preparing the grammar

Specifying rules

PikaParser.epsilon — Constant

epsilon :: Clause

An Epsilon clause. Translate to strongly typed grammar with flatten.

Example

maybe_letter_a = first(token('a'), epsilon)

PikaParser.fail — Constant

fail :: Clause

A Fail clause. Translate to strongly typed grammar with flatten.

Useful for avoiding rule specification when matching terminals using the fast_match parameter of parse.

Example

seq(:this, :that, fail)  # this rule is effectively disabled

PikaParser.first — Method

first(args...) -> PikaParser.First

Build a First clause. Translate to strongly typed grammar with flatten.

Example

first(:something, :fallback, :fallback2)

PikaParser.flatten — Method

flatten(rules::Dict{G}) -> Dict
flatten(rules::Dict{G}, childlabel::Function) -> Dict

Convert a possibly nested and weakly typed rules into a correctly typed and unnested ruleset, usable in make_grammar. This allows use of convenience rule building functions:

satisfy

take_n

token

tokens

epsilon (not a function!)

fail (not a function!)

seq

first

not_followed_by

followed_by

one_or_more

zero_or_more

precedence_cascade (not backed by an actual Clause!)

Anonymous nested rules are assigned names that are constructed by childlabel function (gets the original G and and integer with position integer). By default, childlabel concatenats the parent rule name, hyphen, and the position number to form a Symbol (i.e., the default works only in cases when the rules are labeled by Symbols, and you need to provide your own implementation for other grammars labeled e.g. by integers or strings).

PikaParser.followed_by — Method

followed_by(x) -> PikaParser.FollowedBy

Build a FollowedBy clause. Translate to strongly typed grammar with flatten.

Example

seq(:digits, followed_by(:whitespace))

PikaParser.not_followed_by — Method

not_followed_by(x) -> PikaParser.NotFollowedBy

Build a NotFollowedBy clause. Translate to strongly typed grammar with flatten.

Example

seq(not_followed_by(tokens(collect("reservedWord"))), :identifier)

PikaParser.one_or_more — Method

one_or_more(x) -> PikaParser.OneOrMore

Build a OneOrMore clause. Translate to strongly typed grammar with flatten.

Example

one_or_more(satisfy(isspace))

PikaParser.precedence_cascade — Method

precedence_cascade(label::Function, levels...) -> Vector

Convert a list of rules of increasing associativity to a typical precedence-handling "failthrough" construction. The result must be post-processed by flatten.

Each of the rules is abstracted by "same-associativity" and "higher-associativity" rules (i.e., it is a binary function), which is used to correctly link the rules within the precedence group. The first rule is of the lowest precedence. All rules except the last automatically fallback to the next rule. The higher-precedence parameter of the last rule is the label of the first rule.

label is a function that generates the label for given n-th level of the grammar.

Use @precedences for a less verbose construction.

Returns a vector of labeled rules; that must usually be interpolated into the ruleset.

Example

Dict(
    precedence_cascade(
        n -> Symbol(:exprlevel, n),
        (same, next) -> :expr => first(
            :plus => seq(same, token('+'), next),
            :minus => seq(same, token('-'), next),
        ),
        (same, next) -> :times => seq(same, token('*'), next), # left associative
        (same, next) -> :power => seq(next, token('^'), same), # right associative
        (_, restart) -> first(
            :parens => seq(token('('), restart, token(')')),
            :digits => one_or_more(satisfy(isdigit)),
        ),
    )...,
)

PikaParser.satisfy — Method

satisfy(f::Function) -> PikaParser.Satisfy{Any}

Build a Satisfy clause. Translate to strongly typed grammar with flatten.

Example

satisfy(isdigit)

PikaParser.seq — Method

seq(args...) -> PikaParser.Seq

Build a Seq clause. Translate to strongly typed grammar with flatten.

Example

digit_in_parents = seq(token('('), :digit, token(')'))

PikaParser.take_n — Method

take_n(f::Function) -> PikaParser.TakeN{Any}

Build a TakeN clause. Translate to strongly typed grammar with flatten.

Example

# rule to match a pair of equal tokens
take_n(m -> m[1] == m[2] ? 2 : nothing)

PikaParser.token — Method

token(x) -> PikaParser.Token{Any}

Build a Token clause. Translate to strongly typed grammar with flatten.

Example

token('a')

PikaParser.tokens — Method

tokens(xs::Vector) -> PikaParser.Tokens{Any}

Build a Tokens clause. Translate to strongly typed grammar with flatten.

Example

tokens(collect("keyword"))

PikaParser.zero_or_more — Method

zero_or_more(x) -> PikaParser.ZeroOrMore

Build a ZeroOrMore clause. Translate to strongly typed grammar with flatten.

Example

seq(:quote, zero_or_more(:quote_contents), :quote)

PikaParser.@precedences — Macro

@precedences labeller same::Symbol next::Symbol rules

A shortcut macro for precedence_cascade. Automatically adds lambda heads with fixed argument names, and splats itself with ... into the surrounding environment.

Example

Dict(
    @precedences (n->Symbol(:exprlevel, n)) same next begin
        :expr => seq(same, token('+'), next)
        seq(same, token('*'), next)
        first(
            token('x'),
            seq(token('('), next, token(')'))
        )
    end
)

Converting to a Grammar

PikaParser.make_grammar — Method

make_grammar(starts::AbstractArray{G, 1}, rules_dict::Dict{G, PikaParser.Clause{G}}) -> PikaParser.Grammar

Produce a Grammar with rules of type G that can be used to parse inputs.

starts should collect top-level rules (these will be put at the top of the topological order of the parsing).

rules_dict is a dictionary of grammar Clauses.

Parsing

PikaParser.parse — Function

parse(grammar::PikaParser.Grammar, input::AbstractVector) -> PikaParser.ParseResult
parse(grammar::PikaParser.Grammar, input::AbstractVector, fast_match) -> PikaParser.ParseResult

Take a Grammar and an indexable input sequence, and produce a ParseResult that describes all matched grammar productions.

Fast terminal matching

If fast_match is specified, the function does not match terminals using the associated grammar rules, but with a fast_match function that reports the matched terminals via a callback. The function is called exactly once for each position in input in reverse order (i.e., the indexes will follow reverse(1:length(input)), which can be utilized by the application for optimization). The call parameters consist of the input vector, position in the input vector, and a "report" function used to send back a clause ID (of same type as G in typeof(grammar)) and the length of the terminal matches that can found at that position. Calls to the reporting function can be repeated if more terminal types match. Terminals not reported by the calls to fast_match will not be matched.

For complicated grammars, this may be much faster than having the parser to try matching all terminal types at each position.

Results

Use find_first_parse_at or find_match_at to extract matches from ParseResult.

Pika parsing never really fails. Instead, in case when the grammar rule is not matched in the input, the expected rule match match is either not going to be found at the starting position with find_match_at, or it will not span the whole input.

Example

parse(
    g,
    collect("abcde123"),
    (input, i, match) -> isdigit(input[i]) ? match(:digit, 1) : match(:letter, 1),
)

Traversing and folding the parse tree

PikaParser.find_first_parse_at — Method

find_first_parse_at(grammar::PikaParser.Grammar{G}, parse::PikaParser.ParseResult, pos::Int64) -> Any

Find any possible match of anything starting at input position pos. Preferentially returns the parses that are topologically higher.

If found, returns the Match index in ParseResult, and the corresponding grammar production rule.

PikaParser.find_match_at — Method

find_match_at(grammar::PikaParser.Grammar{G}, parse::PikaParser.ParseResult, rule, pos::Int64) -> Union{Nothing, Int64}

Find the Match index in ParseResult that matched rule at position pos, or nothing if there is no such match.

PikaParser.traverse_match — Method

traverse_match(grammar::PikaParser.Grammar{G}, parse::PikaParser.ParseResult, mid::Int64, rule; open, fold) -> Expr

Given a Match index and the grammar rule matched at that index, recusively depth-first traverse the match tree using functions open (called upon entering a submatch) and fold (called upon leaving the submatch).

open is given the current grammar rule and the UserMatch. It should return a vector of boolean values that tell the traversal which submatches from the UserMatch should be opened. That can be used to skip parsing of large uninteresting parts of the match tree, such as whitespace or comments. By default, it opens the whole subtree.

fold is given the same current grammar rule and the UserMatch, and additionally a vector of folded values from the submatches. The values returned by fold invocations are collected and transferred to higher-level invocations of fold. In case open disabled the evaluation of a given submatch, nothing is used as the folded value for the submatch. By default, fold just collects all submatch values and produces a Julia Expr AST structure where rule expansions are represented as function calls.